Added ImGui, not rendering inside the window atm. Added a few fonts, main menu music. started working on the looting mechanic

+++ IGT - Project/Camera.h
@@ -30,8 +30,8 @@ public:
 	float GetNearDepth() const { return mNearDepth; }
 	float GetFarDepth() const { return mFarDepth; }
 
-	int GetOrthoWidth() const { return mOrthoWidth; }
-	int GetOrthoHeight() const { return mOrthoHeight; }
+	int GetOrthoWidth() const { return (int)mOrthoWidth; }
+	int GetOrthoHeight() const { return (int)mOrthoHeight; }
 
 	Projection GetProjectionMethod() const { return mProjectionMethod; }
