+++ IGT - Project/SoundManager.cpp
@@ -85,7 +85,7 @@ bool SoundManager::PlaySoundEffect(const std::string filename, int channel, int 
 {
 	if (mSoundEffects.find(filename) != mSoundEffects.end())
 	{
-		Mix_Volume(Mix_PlayChannel(channel, mSoundEffects.at(filename), repeat), mSoundEffectsVolume * (mMasterVolume / SDL_MIX_MAXVOLUME));
+		Mix_Volume(Mix_PlayChannel(channel, mSoundEffects.at(filename), repeat), (float)mSoundEffectsVolume * ((float)mMasterVolume / (float)SDL_MIX_MAXVOLUME));
 		return true;
 	}
 	else
@@ -145,7 +145,7 @@ void SoundManager::SetMusicVolume(unsigned int volume)
 	if (volume <= SDL_MIX_MAXVOLUME)
 	{
 		mMusicVolume = volume;
-		Mix_VolumeMusic((int)(mMusicVolume * (mMasterVolume / SDL_MIX_MAXVOLUME)));
+		Mix_VolumeMusic((int)((float)mMusicVolume * (float)((float)mMasterVolume / (float)SDL_MIX_MAXVOLUME)));
 	}
 }
 
@@ -155,6 +155,6 @@ void SoundManager::SetMasterVolume(unsigned int volume)
 	{
 		mMasterVolume = volume;
 
-		Mix_VolumeMusic((int)(mMusicVolume * (mMasterVolume / SDL_MIX_MAXVOLUME)));
+		Mix_VolumeMusic((int)((float)mMusicVolume * (float)((float)mMasterVolume / (float)SDL_MIX_MAXVOLUME)));
 	}
 }
